+++ include/mbgl/map/mode.hpp
@@ -43,7 +43,7 @@ enum class MapDebugOptions : EnumType {
     ParseStatus = 1 << 2,
     Timestamps  = 1 << 3,
     Collision   = 1 << 4,
-    Wireframe   = 1 << 5,
+    Overdraw    = 1 << 5,
 // FIXME: https://github.com/mapbox/mapbox-gl-native/issues/5117
 #ifndef GL_ES_VERSION_2_0
     StencilClip = 1 << 6,