+++ src/mbgl/renderer/painter.cpp
@@ -145,7 +145,7 @@ void Painter::render(const Style& style, const FrameData& frame_, SpriteAtlas& a
         config.depthMask = GL_TRUE;
         config.colorMask = { GL_TRUE, GL_TRUE, GL_TRUE, GL_TRUE };
 
-        if (isWireframe()) {
+        if (isOverdraw()) {
             config.blend = GL_TRUE;
             config.blendFunc = { GL_CONSTANT_COLOR, GL_ONE };
             const float overdraw = 1.0f / 8.0f;
@@ -248,7 +248,7 @@ void Painter::renderPass(RenderPass pass_,
         if (!layer.baseImpl->hasRenderPass(pass))
             continue;
 
-        if (isWireframe()) {
+        if (isOverdraw()) {
             config.blend = GL_TRUE;
         } else if (pass == RenderPass::Translucent) {
             config.blendFunc.reset();