+++ platform/macos/app/MapDocument.m
@@ -276,8 +276,8 @@ NS_ARRAY_OF(id <MGLAnnotation>) *MBXFlattenedShapes(NS_ARRAY_OF(id <MGLAnnotatio
     self.mapView.debugMask ^= MGLMapDebugCollisionBoxesMask;
 }
 
-- (IBAction)toggleWireframes:(id)sender {
-    self.mapView.debugMask ^= MGLMapDebugWireframesMask;
+- (IBAction)toggleOverdrawVisualization:(id)sender {
+    self.mapView.debugMask ^= MGLMapDebugOverdrawVisualizationMask;
 }
 
 - (IBAction)showColorBuffer:(id)sender {
@@ -585,9 +585,9 @@ NS_ARRAY_OF(id <MGLAnnotation>) *MBXFlattenedShapes(NS_ARRAY_OF(id <MGLAnnotatio
         menuItem.title = isShown ? @"Hide Collision Boxes" : @"Show Collision Boxes";
         return YES;
     }
-    if (menuItem.action == @selector(toggleWireframes:)) {
-        BOOL isShown = self.mapView.debugMask & MGLMapDebugWireframesMask;
-        menuItem.title = isShown ? @"Hide Wireframes" : @"Show Wireframes";
+    if (menuItem.action == @selector(toggleOverdrawVisualization:)) {
+        BOOL isShown = self.mapView.debugMask & MGLMapDebugOverdrawVisualizationMask;
+        menuItem.title = isShown ? @"Hide Overdraw Visualization" : @"Show Overdraw Visualization";
         return YES;
     }
     if (menuItem.action == @selector(showColorBuffer:)) {